+++ MoaPlus/Settings/GestureTestView.swift
@@ -686,7 +686,7 @@ private struct SectorOverlay: View {
                     for step in 0...steps {
                         let t = Double(step) / Double(steps)
                         let a = startRad + (endRad - startRad) * t
-                        path.addLine(to: CGPoint(x: cx + radius * cos(a), y: cy + radius * sin(a)))
+                        path.addLine(to: CGPoint(x: cx + radius * CGFloat(cos(a)), y: cy + radius * CGFloat(sin(a))))
                     }
                     path.closeSubpath()
 
@@ -706,8 +706,8 @@ private struct SectorOverlay: View {
                     .font(.system(size: 13, weight: .heavy))
                     .foregroundColor(Color(.label))
                     .position(
-                        x: cx + r * cos(mid * .pi / 180),
-                        y: cy - r * sin(mid * .pi / 180)
+                        x: cx + r * CGFloat(cos(mid * .pi / 180)),
+                        y: cy - r * CGFloat(sin(mid * .pi / 180))
                     )
             }
         }
